Fix a slight bug with the JSRR - JSRR R7.
The previous implementation of JSRR here breaks for JSRR R7. Since R7 was immediately set to PC what happens when you RET is an infinite loop as you execute the subroutine again and again. The modifications here makes it more inline with the LC-3 ISA document as well.

This was a test case that broke my lc3 simulator complx long ago, and I thought I would share a better implementation :)